+++ saas/control_app/src/.aws-sam/build.toml
@@ -1,12 +1,22 @@
 # This file is auto generated by SAM CLI build command
 
-[function_build_definitions.ca87c5f4-6bf1-480c-aeb1-b354b8bb3f09]
-codeuri = "/Users/joecks/git/xayn_terraform_modules/saas/control_app/src/functions"
-runtime = "python3.9"
+[function_build_definitions]
+[function_build_definitions.7f0c1f8a-2100-44a1-ab14-a68a0304ab14]
+codeuri = "/Users/joecks/git/xayn_terraform_modules/saas/control_app/src"
+runtime = "python3.10"
 architecture = "x86_64"
-handler = "authenticator.lambda_handler"
-manifest_hash = "1e8dd909214c4d7a06493a965e0eb4a5"
+handler = "app.functions.authenticator.lambda_handler"
+manifest_hash = "154ebb903b16d5f676852111a7574742"
 packagetype = "Zip"
 functions = ["Authentication"]
+
+[function_build_definitions.72eff071-727a-43c1-8876-3e8e48b1dcca]
+packagetype = "Image"
+functions = ["Provisioning"]
+
+[function_build_definitions.72eff071-727a-43c1-8876-3e8e48b1dcca.metadata]
+DockerTag = "nodejs16_python3.10"
+DockerContext = "/Users/joecks/git/xayn_terraform_modules/saas/control_app/src"
+Dockerfile = "Dockerfile"
 
 [layer_build_definitions]
