+++ saas/control_app/src/.gitignore
@@ -241,4 +241,6 @@ $RECYCLE.BIN/
 
 */build/*
 
+.aws-sam
+
 # End of https://www.gitignore.io/api/osx,linux,python,windows,pycharm,visualstudiocode
