+++ saas/control_app/src/TenantManagement/functions/requirements.txt
@@ -1,0 +1,4 @@
+requests
+boto3
+strenum
+aws-cdk-lib
