+++ saas/control_app/src/requirements.txt
@@ -1,0 +1,1 @@
+-r app/requirements.txt
